n64: specialize compiled blocks
We'd like the recompiler to take the execution context such as kernel
mode into account when compiling blocks. That's why it's necessary to
identify blocks not just by address but all the information used at
compile time. This is done by packing relevant CPU state to an u32 and
storing it alongside the block pointer.

The execution state and its representation as bit vector are recomputed
only when needed, at the moment that means

- each time Context::setMode() is called, happens on powerup,
- on both MTC0 and MFC0 instructions,
- on exceptions,
- when the GDB server requests a block cleanup, and
- when system state is deserialized ("load state").

Recompiler's block pool is now extended with the state bits and a new
compiled block cache. The logic is similar to the old implementation:
a physical address is split into high and low bits, high bits are used
to index a Pool, and the low bits an individual row inside the pool.

The difference: each row represents the most recently used pair of
(Block*, state) instead of just a Block*. The recompiler updates a row
if the state bits don't match by looking up the corresponding block in
the new compiled block cache, which is just unordered_map<u64, Block*>.
This is not supposed to happen often, so even a relatively slow STL
container works fine for the purpose.

Also, there's a new optimization: inlined dual mode instructions.
The recompiler now emits inline implementations of basic 64-bit ops
without runtime kernel mode checks. This is made possible by the block
specialization logic.